requirements.txt
discord.py==1.2.5
loguru
pip
GitPython
discord-timers
python-dateutil